# feat(Makefile): go build optimizations