```
% time make build
make build  129.93s user 11.50s system 348% cpu 40.573 total
```

```
% time make build
make build  8.01s user 0.87s system 133% cpu 6.633 total
```